change in razor pay response

+++ src/main/java/com/LMS/userManagement/Razorpay/OrderRequest.java
@@ -10,6 +10,6 @@ public class OrderRequest {
         String customerName;
         String email;
         String phoneNumber;
-        BigInteger amount;
+        Float amount;
 
 }
